영화추천/requirements.txt
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-cors-headers==2.5.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-imagekit==4.0.2
django-mysql==3.0.0.post1
django-rest-swagger==2.2.0
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pilkit==2.0
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
